añil/U2/FiguresAkinator/pyproject.toml
[tool.poetry]
name = "figuresakinator"
version = "0.1.0"
description = ""
authors = ["luis"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruamel-yaml = "^0.18.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
